requirements.txt
# Telegram Bot API
python-telegram-bot==20.7

# HTTP клиент для асинхронных запросов
aiohttp==3.9.1

# Работа с переменными окружения
python-dotenv==1.0.0

# Логирование
colorlog==6.8.0

# Дополнительные утилиты
asyncio-throttle==1.0.2

# Для работы с JSON и датами
python-dateutil==2.8.2

# Для обработки аудио файлов (если потребуется)
pydub==0.25.1

# Для работы с файлами
pathlib2==2.3.7

